feat(utils.go): add Find function to utils package for finding an element in a slice
The Find function is added to the utils package. It takes a slice of strings and a value to search for. It iterates over the slice and returns the index of the found element and a boolean indicating if the element was found. If the element is not found, it returns -1 and false. This function can be used to search for an element in a slice.